Added hash checks, search engine cross organisation attacks, legend on index.
git-svn-id: https://svn.surfnet.nl/svn/ids/logserver/trunk@291 6e3b1d27-7b68-4cde-b5d6-ae3f64e5b54f

+++ webinterface/report_del.php
@@ -22,12 +22,14 @@ $s_admin = intval($_SESSION['s_admin']);
 $s_userid = intval($_SESSION['s_userid']);
 $s_access = $_SESSION['s_access'];
 $s_access_user = intval($s_access{2});
+$s_hash = md5($_SESSION['s_hash']);
 $err = 0;
 
 $allowed_get = array(
                 "int_userid",
 		"int_rcid",
-		"submit"
+		"submit",
+		"md5_hash"
 );
 $check = extractvars($_GET, $allowed_get);
 debug_input();
@@ -68,38 +70,41 @@ if ($report_content_id > 0) {
   if (pg_num_rows($result_report_content) == 1) {
     $report_content = pg_fetch_assoc($result_report_content);
     
-    # Submit data
-    if (isset($tainted["submit"])) {
-      // First remove refence table
-      if ($report_content["template"] == 3) {
-        // Reference table: report_template_threshold
-        $ref_table = "report_template_threshold";
-        $sql = "DELETE FROM $ref_table WHERE report_content_id = '$report_content_id'";
+    if ($clean['hash'] == $s_hash) {
+      # Submit data
+      if (isset($tainted["submit"])) {
+        // First remove refence table
+        if ($report_content["template"] == 3) {
+          // Reference table: report_template_threshold
+          $ref_table = "report_template_threshold";
+          $sql = "DELETE FROM $ref_table WHERE report_content_id = '$report_content_id'";
+          $debuginfo[] = $sql;
+          $result = pg_query($sql);
+        }            
+        $sql = "DELETE FROM report_content WHERE id = '$report_content_id'";
         $debuginfo[] = $sql;
         $result = pg_query($sql);
-      }            
-      $sql = "DELETE FROM report_content WHERE id = '$report_content_id'";
-      $debuginfo[] = $sql;
-      $result = pg_query($sql);
-      if (pg_affected_rows($result) == 1) {
-        echo "<p style='color:green;'><b>Data succesfully removed.</b></p>\n";
-        echo "<p><a href='mailadmin.php?int_userid=$user_id'>Back</a></p>\n";
-        footer();
-        exit;
-      } else {
-        echo "<p style='color:red;'>Data couldn't be removed (2).</p>\n";
+        if (pg_affected_rows($result) == 1) {
+          echo "<p style='color:green;'><b>Data succesfully removed.</b></p>\n";
+          echo "<p><a href='mailadmin.php?int_userid=$user_id'>Back</a></p>\n";
+          footer();
+          exit;
+        } else {
+          echo "<p style='color:red;'>Data couldn't be removed (2).</p>\n";
+        }
       }
     }
     
     echo "<b>Delete " . $report_content["title"] . "</b><br /><br />\n";
     echo "Are you sure you want to delete this report?<br /><br />\n";
     echo "<form method='get'>\n";
-      echo "<input type='hidden' name='action' value='del'>";
-      echo "<input type='hidden' name='int_userid' value='$user_id'>";
-      echo "<input type='hidden' name='submit' value='1'>";
-      echo "<input type='hidden' name='int_rcid' value='$report_content_id'>";
-      echo "<input type='submit' name='submitBtn' value='Yes' class='button'>\n";
-      echo "<input type='button' name='b1' value='No' onclick=\"window.location.href='mailadmin.php?int_userid=$user_id';\" class='button'>\n";
+      echo "<input type='hidden' name='action' value='del' />";
+      echo "<input type='hidden' name='int_userid' value='$user_id' />";
+      echo "<input type='hidden' name='submit' value='1' />";
+      echo "<input type='hidden' name='int_rcid' value='$report_content_id' />";
+      echo "<input type='hidden' name='md5_hash' value='$s_hash' />\n";
+      echo "<input type='submit' name='submitBtn' value='Yes' class='button' />\n";
+      echo "<input type='button' name='b1' value='No' onclick=\"window.location.href='mailadmin.php?int_userid=$user_id';\" class='button' />\n";
     echo "</form>\n";
   }
 }
